Создан отдельный файл для вспомогательных функций
