+++ src/scripts/utils/utils.js
@@ -1,0 +1,23 @@
+function editDataFormat(wrongDate) {
+  const year = wrongDate.split('-')[0];
+  const month = wrongDate.split('-')[1];
+  const day = wrongDate.split('-')[2];
+  const monthObj = {
+    '01': 'января',
+    '02': 'февраля',
+    '03': 'марта',
+    '04': 'апреля',
+    '05': 'мая',
+    '06': 'июня',
+    '07': 'июля',
+    '08': 'августа',
+    '09': 'сентября',
+    '10': 'октября',
+    '11': 'ноября',
+    '12': 'декабря'
+  };
+  const monthWord = monthObj[`${month}`];
+  return (`${day} ${monthWord}, ${year}`);
+}
+
+export { editDataFormat };
